Contributed by Craig W
* Token Properties now match variables and Default Properties in that they are no longer case sensitive
* Lights/Auras "repainted" correctly from macros where they need to be
* Fix a null pointer error when calling unknown macros on tokens
* getMacroButtonIndex() function corrected to return number
* Ouput from macro calls are formatted correctly if there is only rolls contained in the output.
* Fix problem with json.type on non json objects.
* copyToken accepts a JSON object in optional argument that allows changing of new token values
* raised limit for loops

+++ src/net/rptools/maptool/client/functions/MacroFunctions.java
@@ -1,7 +1,6 @@
 package net.rptools.maptool.client.functions;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedList;
@@ -61,7 +60,7 @@ public class MacroFunctions extends AbstractFunction {
  		} else if (functionName.equals("getMacroCommand")) {
  			return getMacroCommand(resolver, parameters);
 		} else if (functionName.equals("getMacroButtonIndex")) {
-			return BigInteger.valueOf(MapTool.getParser().getMacroButtonIndex());
+			return BigDecimal.valueOf(MapTool.getParser().getMacroButtonIndex());
 		} else if (functionName.equals("removeMacro")) {
 			return removeMacro(resolver, parameters);
 		} else if (functionName.equals("getMacroGroup")) {
